add limit to 16h

+++ supper/views.py
@@ -3,6 +3,8 @@ from django.shortcuts import render, redirect
 from .models import Day, Participation
 from .forms import DayExpenseForm
 from bank.models import Expense, Transaction
+from datetime import datetime as dt
+import datetime
 
 # Create your views here.
 @login_required()
@@ -46,28 +48,37 @@ def day(request, id):
     expensesTotal = 0
     for i in expenses:
         expensesTotal += i.cost
+    canChoose = canChange(id)
     return render(request, 'supper/day.html', locals())
 
 @login_required()
 def switch(request, id):
-    if Participation.objects.filter(user=request.user, day=id).exists(): 
-        Participation.objects.filter(user=request.user, day=id).delete()
-    else:
-        day = Day.objects.get(id=id)
-        Participation.objects.create(day=day, user=request.user)
+    if canChange(id):
+        if Participation.objects.filter(user=request.user, day=id).exists(): 
+            Participation.objects.filter(user=request.user, day=id).delete()
+        else:
+            day = Day.objects.get(id=id)
+            Participation.objects.create(day=day, user=request.user)
     return redirect('supper:day', id=id)
 
 @login_required()
 def upWeight(request, id):
-    p = Participation.objects.get(day=id, user=request.user)
-    p.weight += 1
-    p.save()
+    if canChange(id):
+        p = Participation.objects.get(day=id, user=request.user)
+        p.weight += 1
+        p.save()
     return redirect('supper:day', id=id)
 
 @login_required()
 def downWeight(request, id):
-    p = Participation.objects.get(day=id, user=request.user)
-    if p.weight > 1 :
-        p.weight -= 1
-    p.save()
+    if canChange(id):
+        p = Participation.objects.get(day=id, user=request.user)
+        if p.weight > 1 :
+            p.weight -= 1
+        p.save()
     return redirect('supper:day', id=id)
+
+def canChange(id):
+    day = Day.objects.get(id=id)
+    limit = dt(year=day.date.year, month=day.date.month, day=day.date.day, hour=16)
+    return dt.today() < limit
